WIP: GBTx config converter
So far:
* Added functions to convert registers to values and reverse
* Changed type of reg ID (to be checked for new conversion errors)

Missing:
* Error handling
* Inclusion in HW interface

+++ NSWConfiguration/Types.h
@@ -35,7 +35,7 @@ namespace gbtx {
         std::string description;
         std::size_t size{};
         std::vector<std::size_t> shifts;
-        std::vector<std::size_t> registers;
+        std::vector<std::uint16_t> registers;
     };
 
 }
